refactor(cli): extract render, run and fix from main
Second slice of the cli split. main.py was still 1,545 lines and held the
report renderers, the scan progress loop and the whole fix command alongside
the six other commands.

An AST pass showed the same clean seams as the first slice: the render group
needs nothing at all from the rest of main, and the fix group needs only
_print_report_table and _run_scan.

  render.py   table, HTML, SARIF, badge SVG and the LLM fix plan — pure
              rendering, exercisable on a report object alone
  run.py      _run_scan, the progress loop; its own module because `scan`
              and `fix` both need it and neither should import the other
  fix_cmd.py  the fix command, _run_fix_generation, _run_remote_pr_fix

main.py drops from 1,545 to 864 lines. All 22 top-level functions moved
byte-identical; none were lost, changed or duplicated.

Also stop `--help` ordering from being an accident of import order. Commands
are listed in registration order, so moving one to another module reordered
the help output — as it did here, putting `fix` after `mcp`, and as it did in
the first slice. The order is now stated in one place and applied explicitly,
which removes the whole class of surprise rather than repairing this instance
of it.

Verified beyond unit tests, since the entry point moved again: benchmarks
46/46 taint recall with 0 FP and 3/3 on VAmPI; Juice Shop 198 findings in repo
mode and 26 live DAST findings with an unchanged category split; test-app 114.
The two relocated commands were exercised directly — `fix --help` resolves,
and `badge` renders an SVG through the moved generator.
