+++ isitsecure/cli/main.py
@@ -24,6 +24,14 @@ from isitsecure import __version__
 
 from isitsecure.cli._app import app
 from isitsecure.cli._io import console, err_console
+from isitsecure.cli.render import (
+    _generate_badge_svg,
+    _generate_fixes,
+    _generate_html_report,
+    _generate_sarif_report,
+    _print_report_table,
+)
+from isitsecure.cli.run import _run_scan
 from isitsecure.cli.environment import (
     _LSP_SPECS,
     _chromium_installed,
@@ -495,291 +503,14 @@ def scan(
         raise typer.Exit(1)
 
 
-async def _generate_fixes(report, llm_client, repo_url: str | None) -> str:
-    """Generate LLM-powered fixes for critical and high findings."""
-    from isitsecure.engine.fixes.fix_generator import FixGenerator
-    from isitsecure.engine.fixes.markdown_exporter import FixPlanMarkdownExporter
-
-    # Filter to fixable findings (SAST with code locations)
-    fixable = [
-        f for f in report.findings
-        if f.code_location and f.code_location.file_path
-        and f.severity.value in ("critical", "high")
-    ]
-
-    if not fixable:
-        return "# isitsecure Fix Plan\n\nNo critical or high findings with source code locations to fix."
-
-    # Build file content map from the report's findings
-    # If we have a local repo, read the files directly
-    file_contents: dict[str, str] = {}
-    if repo_url and repo_url.startswith("file://"):
-        import os
-        repo_path = repo_url.replace("file://", "").rstrip("/")
-        for finding in fixable:
-            fp = finding.code_location.file_path
-            if fp not in file_contents:
-                full_path = os.path.join(repo_path, fp)
-                if os.path.isfile(full_path):
-                    try:
-                        file_contents[fp] = open(full_path).read()
-                    except Exception:
-                        pass
-
-    # Fall back to code snippets from findings if we can't read files
-    for finding in fixable:
-        fp = finding.code_location.file_path
-        if fp not in file_contents and finding.code_location.code_snippet:
-            file_contents[fp] = finding.code_location.code_snippet
-
-    console.print(f"  Generating fixes for {len(fixable)} findings across {len(file_contents)} files...")
-
-    generator = FixGenerator(llm_client)
-    plan = await generator.generate_fix_plan(fixable, file_contents)
-
-    exporter = FixPlanMarkdownExporter()
-    return exporter.export(plan)
 
 
-def _generate_sarif_report(report) -> str:
-    """Generate a SARIF 2.1.0 report from a DeepScanReport."""
-    from isitsecure.engine.reporting.sarif_renderer import SARIFRenderer
-
-    renderer = SARIFRenderer()
-    return renderer.render(report)
 
 
-def _generate_html_report(report) -> str:
-    """Generate a self-contained HTML report from a DeepScanReport."""
-    from isitsecure.engine.reporting.report_generator import ReportGenerator
-    from isitsecure.engine.reporting.html_renderer import HTMLReportRenderer
-
-    generator = ReportGenerator()
-    renderer = HTMLReportRenderer()
-    report_data = generator.generate(report)
-    return renderer.render(report_data)
 
 
-async def _run_scan(agent, **kwargs):
-    """Run the scan, narrating each step as a live scrolling log.
-
-    The tool "speaks" what it's doing — a phase header for each stage and an
-    indented line as each scanner finishes — so a long scan visibly progresses
-    instead of freezing on a single bar.
-    """
-    import time
-
-    report = None
-    failure: str | None = None
-    t0 = time.monotonic()
-    last_phase = None
-    err_console.print()
-
-    async for event in agent.scan(**kwargs):
-        phase = getattr(event, "phase", "")
-        phase_val = getattr(phase, "value", phase)
-        message = getattr(event, "message", "") or "Scanning..."
-        data = getattr(event, "data", None) or {}
-        elapsed = time.monotonic() - t0
-        stamp = f"[dim]{elapsed:6.1f}s[/dim]"
-
-        # The final COMPLETE event carries the report; capture it, don't log it.
-        if "report" in data:
-            from isitsecure.engine.models import DeepScanReport
-            report = DeepScanReport.model_validate(data["report"])
-            continue
-
-        # Keep the reason a phase gave up, so a scan that ends without a
-        # report can explain itself instead of shrugging (#147).
-        if data.get("error") and message:
-            failure = message
-
-        status = data.get("status")
-        if status == "start":
-            # Scanner launched — show it's in flight.
-            err_console.print(f"{stamp}    [cyan]→[/cyan] [dim]{data['scanner']}…[/dim]")
-        elif status == "done":
-            # Scanner finished — detail line.
-            count = data.get("findings", 0)
-            if count:
-                err_console.print(
-                    f"{stamp}    [green]✓[/green] {data['scanner']} "
-                    f"[yellow]— {count} finding(s)[/yellow]"
-                )
-            else:
-                err_console.print(
-                    f"{stamp}    [green]✓[/green] [dim]{data['scanner']} — clean[/dim]"
-                )
-        elif phase_val != last_phase:
-            # New phase — header line.
-            err_console.print(f"{stamp} [bold cyan]▶[/bold cyan] {message}")
-            last_phase = phase_val
-        else:
-            # A sub-step within the current phase (emitted by a scanner).
-            err_console.print(f"{stamp}      [dim]· {message}[/dim]")
-
-    if report is None:
-        if failure:
-            err_console.print(f"\n[red]{failure}[/red]")
-        else:
-            err_console.print(
-                "[red]The scan finished but didn't produce any results — something "
-                "went wrong along the way.[/red]"
-            )
-        err_console.print(
-            "[dim]Try re-running with -v (verbose) to see what happened, or check "
-            "that your website address / code path is correct.[/dim]"
-        )
-        raise typer.Exit(1)
-
-    err_console.print(
-        f"[dim]{time.monotonic() - t0:6.1f}s[/dim] [green]✓ Scan complete[/green]\n"
-    )
-
-    return report
 
 
-def _print_report_table(report) -> None:
-    """Print a summary table of the scan report.
-
-    Leads with a rule-based, LLM-free launch-readiness verdict and a
-    granular grade so a non-technical user gets a clear go/no-go up top,
-    then a business-impact-first findings table with plain-English framing.
-    """
-    from isitsecure.engine.reporting import plain_english
-
-    # #43 — granular grade (A+/A/A-/.../F) + plain-language legend.
-    grade_result = plain_english.calculate_grade(
-        critical=report.critical_count,
-        high=report.high_count,
-        medium=report.medium_count,
-        low=sum(
-            1 for f in report.findings
-            if (f.severity.value if hasattr(f.severity, "value") else f.severity)
-            == "low"
-        ),
-    )
-    grade = grade_result.grade
-
-    # #57 — go/no-go launch verdict, rendered first and most prominently.
-    verdict = plain_english.launch_verdict(
-        report.critical_count, report.high_count, report.medium_count
-    )
-    console.print()
-    console.print(Panel(
-        f"[bold]{verdict.headline}[/bold]"
-        + (f"\n{verdict.detail}" if verdict.detail else ""),
-        title="Launch Readiness",
-        border_style="green" if verdict.ready else "red",
-    ))
-
-    console.print(Panel(
-        f"[bold]Grade: {grade}[/bold] — {grade_result.label}\n"
-        f"[dim]{grade_result.legend}[/dim]\n\n"
-        f"Critical: {report.critical_count}  |  "
-        f"High: {report.high_count}  |  "
-        f"Medium: {report.medium_count}  |  "
-        f"Endpoints: {report.total_endpoints_discovered}  |  "
-        f"Scanners: {len(report.scanners_run)}  |  "
-        f"Duration: {report.scan_duration_seconds:.0f}s",
-        title="Results",
-        border_style="bright_magenta",
-    ))
-
-    if not report.findings:
-        console.print("[green]No vulnerabilities found![/green]")
-        return
-
-    # Findings table — business-impact-first (#44), with plain-English
-    # framing and inline glossary (#41, #42).
-    table = Table(title="Findings", show_lines=True)
-    table.add_column("#", style="dim", width=3)
-    table.add_column("Severity", width=9)
-    table.add_column("What this means for you", width=48)
-    table.add_column("Category", width=18)
-    table.add_column("Detail", width=40)
-
-    severity_colors = {
-        "critical": "red bold",
-        "high": "red",
-        "medium": "yellow",
-        "low": "blue",
-        "info": "dim",
-    }
-
-    # Order most-severe first so the biggest risks are read first.
-    severity_order = {
-        "critical": 0, "high": 1, "medium": 2, "low": 3, "info": 4,
-    }
-
-    def _sev(f):
-        return f.severity.value if hasattr(f.severity, "value") else str(f.severity)
-
-    ordered = sorted(report.findings, key=lambda f: severity_order.get(_sev(f), 5))
-
-    seen_glossary: set[str] = set()
-    for i, finding in enumerate(ordered, 1):
-        sev = _sev(finding)
-        color = severity_colors.get(sev, "white")
-        category = finding.category.value if hasattr(finding.category, "value") else str(finding.category)
-
-        # #44 — consequence-first summary column.
-        impact = plain_english.business_impact(finding.category)
-        # #42 — expand each acronym once on first use (parenthetical).
-        detail = finding.title[:40]
-        for term, definition in plain_english.GLOSSARY.items():
-            if term in seen_glossary:
-                continue
-            import re as _re
-            if _re.search(rf"\b{_re.escape(term)}\b", f"{finding.title} {category}".lower()):
-                detail = f"{detail}\n[dim]{term.upper()}: {definition}[/dim]"
-                seen_glossary.add(term)
-                break
-
-        # Stable fingerprint so the user can `--suppress` this finding (#38/#51).
-        detail = f"{detail}\n[dim]fp {finding.fingerprint}[/dim]"
-
-        table.add_row(
-            str(i),
-            f"[{color}]{sev.upper()}[/{color}]",
-            impact,
-            category,
-            detail,
-        )
-
-    console.print(table)
-
-    # #49 — step-by-step walkthroughs for the top-4 fixes, as numbered lists.
-    # One walkthrough per category present (deduped), so a repo with several
-    # IDOR findings shows the "add an ownership check" steps once.
-    seen_walkthroughs: set[str] = set()
-    for finding in ordered:
-        walkthrough = plain_english.walkthrough_for(finding.category)
-        if walkthrough is None:
-            continue
-        cat = finding.category.value if hasattr(finding.category, "value") else str(finding.category)
-        if cat in seen_walkthroughs:
-            continue
-        seen_walkthroughs.add(cat)
-        steps = "\n".join(
-            f"[bold]{i}.[/bold] {step}"
-            for i, step in enumerate(walkthrough.steps, 1)
-        )
-        console.print()
-        console.print(Panel(
-            steps,
-            title=f"How to fix, step by step: {walkthrough.title}",
-            border_style="cyan",
-        ))
-
-    # Owner summary (LLM layer, if present) — layers on top of the baseline.
-    if report.owner_summary and report.owner_summary.risk_summary:
-        console.print()
-        console.print(Panel(
-            report.owner_summary.risk_summary,
-            title="Risk Summary",
-            border_style="yellow",
-        ))
 
 
 # ---------------------------------------------------------------------------
@@ -919,379 +650,10 @@ def launch(
 # fix command — scan + generate fixes + apply them
 # ---------------------------------------------------------------------------
 
-@app.command()
-def fix(
-    repo: str = typer.Option(..., "--repo", "-r", help="Local repo path, OR a remote GitHub URL to open PRs against"),
-    llm_provider: str = typer.Option("anthropic", "--llm", help="LLM provider: anthropic|google"),
-    api_key: Optional[str] = typer.Option(None, "--api-key", envvar="ANTHROPIC_API_KEY"),
-    dry_run: bool = typer.Option(False, "--dry-run", help="Show fixes without applying them"),
-    severity: str = typer.Option("critical,high", "--severity", help="Severities to fix: critical,high,medium"),
-    technical: bool = typer.Option(
-        False, "--technical",
-        help="Show the git details (backup ref, diff/test commands) instead of the plain-language summary",
-    ),
-    github_token: Optional[str] = typer.Option(None, "--github-token", envvar="GITHUB_TOKEN", help="GitHub token for remote-repo pull requests (never stored/logged)"),
-    pr_strategy: str = typer.Option("per-category", "--pr-strategy", help="Group PRs by: per-category|per-file|per-finding|single"),
-    max_prs: int = typer.Option(8, "--max-prs", help="Cap on PRs; excess low-severity categories batch into one PR"),
-    verbose: bool = typer.Option(False, "--verbose", "-v"),
-) -> None:
-    """Scan your code, fix the issues, and re-check — in plain language.
-
-    A local ``--repo`` path gets fixes applied in place, git-free by default —
-    your original is safely backed up under the hood; pass --technical for the
-    git details. A remote GitHub ``--repo`` URL (with ``--github-token``) is
-    cloned and gets per-category pull requests opened instead.
-    """
-    if verbose:
-        logging.basicConfig(level=logging.DEBUG)
-    else:
-        logging.basicConfig(level=logging.WARNING)
-
-    # Resolve API key
-    resolved_key = api_key or _load_api_key(llm_provider)
-    if not resolved_key:
-        err_console.print(
-            "[red]Auto-fixing your code needs the AI turned on, and I couldn't "
-            "find an API key.[/red]\n"
-            "[bold]Fix:[/bold] run [dim]isitsecure setup[/dim] to add one "
-            f"(or set {llm_provider.upper()}_API_KEY, or pass --api-key)."
-        )
-        raise typer.Exit(1)
-
-    from isitsecure.llm.adapters import create_llm_client
-    llm_client = create_llm_client(llm_provider, resolved_key)
-
-    # Remote GitHub URL → clone + per-category pull requests.
-    from isitsecure.engine.fixes.pr_flow import is_remote_url
-    if is_remote_url(repo):
-        _run_remote_pr_fix(
-            repo_url=repo,
-            llm_client=llm_client,
-            llm_provider=llm_provider,
-            github_token=github_token,
-            severity=severity,
-            pr_strategy=pr_strategy,
-            max_prs=max_prs,
-            dry_run=dry_run,
-        )
-        return
-
-    # Resolve repo path
-    import os
-    repo_path = os.path.abspath(repo.replace("file://", ""))
-    if not os.path.isdir(repo_path):
-        err_console.print(
-            f"[red]I couldn't find your code at:[/red] {repo_path}\n"
-            "[dim]Double-check the path — --repo should point to a folder on your "
-            "computer (e.g. --repo ./my-app or --repo /Users/you/my-app).[/dim]"
-        )
-        raise typer.Exit(1)
-
-    repo_url = f"file://{repo_path}"
-
-    # Step 1: Scan
-    console.print(Panel(
-        f"[bold]isitsecure fix[/bold]\n"
-        f"Repo: {repo_path}  |  LLM: {llm_provider}  |  {'Dry run' if dry_run else 'Will apply fixes'}",
-        title="Auto-Fix",
-        border_style="bright_magenta",
-    ))
-
-    from isitsecure.engine.factory import create_deep_security_scan_agent, create_repo_ingestion_service
-    repo_service = create_repo_ingestion_service()
-    agent = create_deep_security_scan_agent(
-        llm_client=llm_client,
-        judgment_llm_client=llm_client,
-        repo_ingestion_service=repo_service,
-    )
-
-    console.print("\n[bold]Step 1/3:[/bold] Scanning for vulnerabilities...")
-    report = asyncio.run(_run_scan(agent=agent, repo_url=repo_url, scan_mode=None))
-    _print_report_table(report)
-
-    # Step 2: Generate fixes
-    target_severities = {s.strip().lower() for s in severity.split(",")}
-    fixable = [
-        f for f in report.findings
-        if f.code_location and f.code_location.file_path
-        and (f.severity.value if hasattr(f.severity, "value") else str(f.severity)) in target_severities
-    ]
-
-    if not fixable:
-        console.print("\n[green]No fixable findings at the selected severity levels.[/green]")
-        raise typer.Exit(0)
-
-    console.print(f"\n[bold]Step 2/3:[/bold] Generating fixes for {len(fixable)} findings...")
-
-    # Read file contents
-    file_contents: dict[str, str] = {}
-    for finding in fixable:
-        fp = finding.code_location.file_path
-        if fp not in file_contents:
-            full_path = os.path.join(repo_path, fp)
-            if os.path.isfile(full_path):
-                try:
-                    file_contents[fp] = open(full_path).read()
-                except Exception:
-                    pass
-
-    fix_plan = asyncio.run(_run_fix_generation(llm_client, fixable, file_contents))
-
-    if not fix_plan.files:
-        console.print("\n[yellow]No fixes could be generated.[/yellow]")
-        if fix_plan.skipped:
-            for reason in fix_plan.skipped:
-                console.print(f"  [dim]Skipped: {reason}[/dim]")
-        raise typer.Exit(0)
-
-    # Step 3: Apply fixes — one final version per file. Multiple findings in
-    # the same file are chained into a single rewrite (no clobbering).
-    from difflib import unified_diff
-    n_files = len(fix_plan.files)
-
-    # --- Dry run: just preview the diffs, change nothing. ---
-    if dry_run:
-        console.print(
-            f"\n[bold]Step 3/3:[/bold] Previewing fixes for {fix_plan.fixed_count} "
-            f"findings across {n_files} file(s)..."
-        )
-        for path, fixed_content in fix_plan.files.items():
-            console.print(f"\n  [bold]{path}[/bold]")
-            original = file_contents.get(path, "")
-            diff = "\n".join(unified_diff(
-                original.splitlines(), fixed_content.splitlines(),
-                fromfile=f"a/{path}", tofile=f"b/{path}", lineterm="",
-            ))
-            console.print(f"  [dim]{diff[:800]}[/dim]")
-        console.print(
-            f"\n[dim]Run without --dry-run to apply these fixes: "
-            f"isitsecure fix --repo {repo}[/dim]"
-        )
-        raise typer.Exit(0)
-
-    # --- Apply for real: take a safety net first, then write files in place. ---
-    console.print(
-        f"\n[bold]Step 3/3:[/bold] Applying fixes for {fix_plan.fixed_count} "
-        f"findings across {n_files} file(s)..."
-    )
-
-    from isitsecure.engine.fixes.safety_net import create_safety_net
-    from isitsecure.engine.shared.safe_path import resolve_within
-
-    net = create_safety_net(repo_path, list(fix_plan.files.keys()))
-
-    applied = 0
-    failed = 0
-    for path, fixed_content in fix_plan.files.items():
-        try:
-            full_path = resolve_within(repo_path, path)
-            with open(full_path, "w") as f:
-                f.write(fixed_content)
-            applied += 1
-        except Exception as e:
-            console.print(f"  [red]Couldn't update {path}: {e}[/red]")
-            failed += 1
-
-    # Re-scan the fixed code to confirm the findings are actually gone.
-    from isitsecure.engine.fixes.verifier import verify_findings_resolved
-    from isitsecure.engine.fixes import plain_results
-
-    fixed_findings = [
-        f for f in fixable
-        if f.code_location and f.code_location.file_path in fix_plan.files
-    ]
-    console.print("[bold]Re-checking your code...[/bold]")
-    vr = asyncio.run(verify_findings_resolved(repo_path, fixed_findings))
-
-    # Fold everything into the three plain-language buckets. "couldn't fix" =
-    # findings we tried but produced no fix for (failed generation + write
-    # failures).
-    fix_failed = (len(fixable) - fix_plan.fixed_count) + failed
-    counts = plain_results.classify_verification(
-        attempted=len(fixable),
-        fix_failed=fix_failed,
-        verification=vr.to_dict(),
-    )
-
-    console.print()
-    console.print(Panel(
-        f"[bold]{plain_results.summarize(counts)}[/bold]",
-        title="Done",
-        border_style="green" if counts.needs_review == 0 and counts.couldnt_fix == 0 else "yellow",
-    ))
-
-    hint = plain_results.next_step_hint(counts, saved_hint=net.restore_hint)
-    if hint:
-        console.print(f"\n[dim]{hint}[/dim]")
-
-    if counts.needs_review and vr.still_present_titles:
-        console.print("\n[bold]Worth a look:[/bold]")
-        for t in vr.still_present_titles:
-            console.print(f"  [yellow]•[/yellow] {t}")
-
-    # --- Power-user / technical view: the git mechanics, on request. ---
-    if technical:
-        console.print("\n[bold]Technical details:[/bold]")
-        if net.kind == "git":
-            console.print(
-                f"  Backup ref: [dim]{net.location}[/dim] "
-                f"[dim](restore original: git checkout {net.location} -- .)[/dim]"
-            )
-        elif net.kind == "copy":
-            console.print(
-                f"  Backup copy: [dim]{net.location}[/dim] "
-                f"[dim](restore original: cp -a {net.location}/. .)[/dim]"
-            )
-        console.print("  1. Review changes: [dim]git diff[/dim]")
-        console.print("  2. Run your tests")
-        console.print(
-            "  3. Add isitsecure to CI so it can't regress "
-            "([dim]see examples/github-action.yml[/dim])"
-        )
 
 
-async def _run_fix_generation(llm_client, findings, file_contents):
-    """Run fix generation, chaining multiple findings per file (no clobbering)."""
-    from isitsecure.engine.fixes.fix_generator import FixGenerator
-
-    generator = FixGenerator(llm_client)
-    return await generator.generate_file_fixes(findings, file_contents)
 
 
-def _run_remote_pr_fix(
-    *,
-    repo_url: str,
-    llm_client,
-    llm_provider: str,
-    github_token: Optional[str],
-    severity: str,
-    pr_strategy: str,
-    max_prs: int,
-    dry_run: bool,
-) -> None:
-    """Scan a REMOTE GitHub repo and open per-category pull requests with fixes."""
-    from isitsecure.engine.fixes.pr_flow import (
-        parse_github_url,
-        group_findings,
-        PRFlow,
-    )
-
-    # Parse + guard the host up front so we fail clearly before scanning.
-    try:
-        ref = parse_github_url(repo_url)
-    except ValueError as exc:
-        err_console.print(f"[red]Couldn't parse that repo URL:[/red] {exc}")
-        raise typer.Exit(1)
-
-    if not ref.is_github:
-        err_console.print(
-            f"[yellow]{ref.host} pull requests aren't supported yet — GitHub only.[/yellow]\n"
-            "[dim]Scan locally and run 'isitsecure fix --repo <path>' to apply fixes in place.[/dim]"
-        )
-        raise typer.Exit(1)
-
-    if not github_token:
-        err_console.print(
-            "[red]Opening pull requests on a remote repo needs a GitHub token.[/red]\n"
-            "[bold]Fix:[/bold] pass [dim]--github-token[/dim] (or set GITHUB_TOKEN). "
-            "It's used only for the push + PR and is never stored or logged."
-        )
-        raise typer.Exit(1)
-
-    console.print(Panel(
-        f"[bold]isitsecure fix (remote)[/bold]\n"
-        f"Repo: {ref.slug}  |  LLM: {llm_provider}  |  Strategy: {pr_strategy}  |  Max PRs: {max_prs}",
-        title="Remote Auto-Fix → Pull Requests",
-        border_style="bright_magenta",
-    ))
-
-    # Step 1: Scan the remote repo (SAST — code-only).
-    from isitsecure.engine.factory import (
-        create_deep_security_scan_agent,
-        create_repo_ingestion_service,
-    )
-    repo_service = create_repo_ingestion_service()
-    agent = create_deep_security_scan_agent(
-        llm_client=llm_client,
-        judgment_llm_client=llm_client,
-        repo_ingestion_service=repo_service,
-    )
-
-    console.print("\n[bold]Step 1/2:[/bold] Scanning the remote repo for vulnerabilities...")
-    from isitsecure.engine.enums import ScanMode
-    report = asyncio.run(
-        _run_scan(agent=agent, repo_url=repo_url, scan_mode=ScanMode.CODE_ONLY)
-    )
-    _print_report_table(report)
-
-    target_severities = {s.strip().lower() for s in severity.split(",")}
-    fixable = [
-        f for f in report.findings
-        if f.code_location and f.code_location.file_path
-        and (f.severity.value if hasattr(f.severity, "value") else str(f.severity)) in target_severities
-    ]
-    if not fixable:
-        console.print("\n[green]No fixable findings at the selected severity levels.[/green]")
-        raise typer.Exit(0)
-
-    if dry_run:
-        # Show the PR plan without cloning/pushing.
-        groups = group_findings(fixable, strategy=pr_strategy, max_prs=max_prs)
-        console.print(f"\n[bold]Dry run — would open {len(groups)} pull request(s):[/bold]")
-        for g in groups:
-            label = "low-severity cleanup" if g.is_low_batch else g.title_label
-            console.print(f"  • [cyan]isitsecure/fix-{g.branch_suffix}[/cyan] — {label} "
-                          f"({len(g.findings)} finding(s))")
-        console.print("\n[dim]Run without --dry-run to clone, push branches, and open the PRs.[/dim]")
-        raise typer.Exit(0)
-
-    # Step 2: Clone → fix → group → push → open PRs.
-    console.print(f"\n[bold]Step 2/2:[/bold] Generating fixes and opening pull requests...")
-
-    from isitsecure.engine.fixes.fix_generator import FixGenerator
-
-    async def _emit(event: dict) -> None:
-        msg = event.get("message", "")
-        if msg:
-            console.print(f"  [dim]{msg}[/dim]")
-
-    async def _go():
-        flow = PRFlow(FixGenerator(llm_client))
-        return await flow.run(
-            repo_url=repo_url,
-            findings=fixable,
-            github_token=github_token,
-            strategy=pr_strategy,
-            max_prs=max_prs,
-            emit=_emit,
-        )
-
-    try:
-        result = asyncio.run(_go())
-    except ValueError as exc:
-        err_console.print(f"[red]{exc}[/red]")
-        raise typer.Exit(1)
-
-    # Report
-    console.print()
-    if result.opened_prs:
-        console.print("[bold green]Pull requests opened:[/bold green]")
-        for pr in result.opened_prs:
-            console.print(f"  • {pr.title}")
-            console.print(f"    [cyan]{pr.url}[/cyan]")
-    console.print(Panel(
-        f"[bold]{result.summary}[/bold]\n"
-        f"{result.fixed_count} finding(s) fixed  |  "
-        f"{len(result.skipped)} skipped  |  {len(result.errors)} error(s)",
-        title="Remote Fix Summary",
-        border_style="green" if not result.errors else "yellow",
-    ))
-    if result.errors:
-        for e in result.errors:
-            console.print(f"  [yellow]• {e}[/yellow]")
-    if result.skipped and logging.getLogger().isEnabledFor(logging.DEBUG):
-        for s in result.skipped:
-            console.print(f"  [dim]Skipped: {s}[/dim]")
 
 
 # ---------------------------------------------------------------------------
@@ -1342,48 +704,6 @@ def badge(
     console.print(f'  [dim]![Security: {grade}]({badge_ref})[/dim]')
 
 
-def _generate_badge_svg(grade: str, critical: int, high: int, total: int) -> str:
-    """Generate a shields.io-style SVG badge for the security grade."""
-    GRADE_COLORS = {
-        "A": "#4c1",      # Green
-        "B": "#97ca00",   # Yellow-green
-        "C": "#dfb317",   # Yellow
-        "D": "#fe7d37",   # Orange
-        "F": "#e05d44",   # Red
-    }
-    # Grades are now granular (A+, A-, C+, ...); color by the base letter.
-    color = GRADE_COLORS.get(grade[:1], "#9f9f9f")
-
-    label = "security"
-    value = grade
-    if total > 0:
-        value = f"{grade} ({total} findings)"
-
-    label_width = len(label) * 6.5 + 10
-    value_width = len(value) * 6.5 + 10
-    total_width = label_width + value_width
-
-    return f'''<svg xmlns="http://www.w3.org/2000/svg" width="{total_width}" height="20" role="img" aria-label="{label}: {value}">
-  <title>{label}: {value}</title>
-  <linearGradient id="s" x2="0" y2="100%">
-    <stop offset="0" stop-color="#bbb" stop-opacity=".1"/>
-    <stop offset="1" stop-opacity=".1"/>
-  </linearGradient>
-  <clipPath id="r">
-    <rect width="{total_width}" height="20" rx="3" fill="#fff"/>
-  </clipPath>
-  <g clip-path="url(#r)">
-    <rect width="{label_width}" height="20" fill="#555"/>
-    <rect x="{label_width}" width="{value_width}" height="20" fill="{color}"/>
-    <rect width="{total_width}" height="20" fill="url(#s)"/>
-  </g>
-  <g fill="#fff" text-anchor="middle" font-family="Verdana,Geneva,DejaVu Sans,sans-serif" text-rendering="geometricPrecision" font-size="11">
-    <text aria-hidden="true" x="{label_width/2}" y="15" fill="#010101" fill-opacity=".3">{label}</text>
-    <text x="{label_width/2}" y="14">{label}</text>
-    <text aria-hidden="true" x="{label_width + value_width/2}" y="15" fill="#010101" fill-opacity=".3">{value}</text>
-    <text x="{label_width + value_width/2}" y="14">{value}</text>
-  </g>
-</svg>'''
 
 
 # ---------------------------------------------------------------------------
